requirements.txt
fastapi
uvicorn[standard]
python-multipart
faster-whisper
openai-whisper
pydantic-settings
requests
